モーダルの中身をクリックした制作物の情報になるよう修正
